refactor : 기존 Member의 Command와 Query 패키지에 가각 존재하던 MemberProfileFacade를 orgstructure하위로 통합 편성
